research-agent/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
arxiv==2.1.0
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.2.0
beautifulsoup4==4.12.3
lxml==5.2.2